refactor. step5 - 클래스 디렉토리 변경 및 이동
