+++ src/main/java/basicgame/domain/CarTryCounter.java
@@ -1,4 +1,4 @@
-package basicgame;
+package basicgame.domain;
 
 public class CarTryCounter {
