+++ src/main/java/basicgame/service/strategy/ActiveStrategy.java
@@ -1,4 +1,4 @@
-package basicgame.strategy;
+package basicgame.service.strategy;
 
 public interface ActiveStrategy {
     int generateRandom(int max);
