+++ src/main/java/basicgame/service/strategy/RandomActive.java
@@ -1,4 +1,4 @@
-package basicgame.strategy;
+package basicgame.service.strategy;
 
 import util.RandomUtil;
