+++ src/main/java/basicgame/service/strategy/RandomActiveMock.java
@@ -1,4 +1,4 @@
-package basicgame.strategy;
+package basicgame.service.strategy;
 
 public class RandomActiveMock implements ActiveStrategy{
